Move back to own directory
