+++ frontend/assets/data/defaultSettings.tsx
