+++ frontend/assets/entities/configuration.entity.ts
