+++ frontend/assets/jss/components/card.tsx
