+++ frontend/assets/jss/shared.tsx
