+++ frontend/assets/jss/tooltips.tsx
