+++ frontend/components/Data/Data.tsx
@@ -14,7 +14,7 @@ import {
 import { Event } from "../../assets/entities/event.entity";
 import { useSettings } from "../Common/Utils";
 import { WebSocketConnection } from "../Common/WebSocket";
-import DataItems from "../Data/DataItems";
+import DataItems from "./DataItems";
 import Footer from "../Common/Footer";
 
 const items = [
