+++ frontend/components/Logs/Logs.tsx
