+++ frontend/next.config.js
@@ -1,5 +1,4 @@
 module.exports = {
-  distDir: "../../../.next",
   eslint: {
     ignoreDuringBuilds: true,
   },
