+++ frontend/package.json
@@ -1,47 +1,43 @@
 {
   "name": "system-bridge-frontend",
+<<<<<<< HEAD
   "version": "2.6.0",
+=======
+  "version": "1.0.0",
+>>>>>>> ec30169 (Move back to own directory)
   "private": true,
   "scripts": {
     "build": "next build",
     "eslint": "eslint --ext .ts --ext .tsx --color ./",
-    "start:dev": "next dev",
-    "start": "next start"
+    "start": "next start",
+    "start:dev": "next dev"
   },
   "dependencies": {
+    "@fontsource/roboto": "4.5.0",
     "@material-ui/core": "4.12.3",
     "@material-ui/icons": "4.11.2",
     "@material-ui/lab": "4.0.0-alpha.60",
     "@material-ui/styles": "4.11.4",
     "@mdi/js": "5.9.55",
     "@mdi/react": "1.5.0",
-    "@types/uuid": "^8.3.1",
-    "axios": "0.21.1",
     "clsx": "1.1.1",
-    "fontsource-roboto": "4.0.0",
     "lodash": "4.17.21",
-    "moment": "2.29.1",
     "next": "11.1.0",
     "peerjs": "1.3.2",
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "react-json-view": "1.21.3",
-    "react-router-dom": "5.2.1",
-    "react-scripts": "4.0.3",
-    "rect-scaler": "1.0.0",
-    "uuid": "8.3.2"
+    "react-markdown": "7.0.1",
+    "react-moment": "1.1.1",
+    "react-syntax-highlighter": "15.4.4",
+    "rect-scaler": "1.0.0"
   },
   "license": "MIT",
   "devDependencies": {
-    "@types/node": "16.7.10",
     "@types/react": "17.0.19",
+    "@types/react-dom": "17.0.9",
     "@types/react-syntax-highlighter": "13.5.2",
-    "@typescript-eslint/eslint-plugin": "4.29.3",
-    "@typescript-eslint/parser": "4.29.3",
-    "eslint": "7.32.0",
     "eslint-import-resolver-typescript": "2.4.0",
-    "eslint-plugin-import": "2.24.2",
-    "prettier": "2.3.2",
-    "typescript": "4.4.2"
+    "eslint-plugin-import": "2.24.2"
   }
 }
