+++ frontend/pages/app/logs.tsx
