+++ package.json
@@ -19,15 +19,19 @@
   "license": "MIT",
   "bin": "dist/main.js",
   "scripts": {
-    "build": "yarn build:app && yarn build:tray",
-    "build:app": "rimraf dist && nest build && cd src/components/frontend && next build",
+    "build": "yarn build:app && yarn build:frontend && yarn build:tray",
+    "build:app": "rimraf dist && nest build",
+    "build:frontend": "cd frontend && yarn build",
     "build:tray": "cd tray && yarn build",
     "format": "prettier --write \"src/**/*.ts\" \"test/**/*.ts\"",
-    "install:all": "yarn install && cd tray && yarn install",
-    "install:all:ci": "yarn install --pure-lockfile && cd tray && yarn install --pure-lockfile",
+    "install:all": "yarn install && cd frontend && yarn install && cd ../tray && yarn install",
+    "install:all:ci": "yarn install --pure-lockfile && cd frontend && yarn install --pure-lockfile && cd ../tray && yarn install --pure-lockfile",
     "lint": "eslint \"{src,apps,libs,test}/**/*.ts\" --fix",
     "package": "rimraf out && yarn build && node scripts/package.js",
     "package:app": "rimraf out && yarn build:app && cross-env APP_ONLY=true node scripts/package.js",
+    "start": "concurrently -k \"yarn start:app\" \"yarn start:frontend\" \"yarn start:tray\"",
+    "start:dev": "concurrently -k \"yarn start:app:dev\" \"yarn start:frontend:dev\" \"yarn start:tray\"",
+    "start:prod": "concurrently -k \"yarn start:app:prod\" \"yarn start:tray:prod\"",
     "start:app": "cross-env NODE_ENV=development nest start",
     "start:app:cli": "cross-env NODE_ENV=development CLI_ONLY=true SB_TRAY=false nest start",
     "start:app:cli:dev": "cross-env NODE_ENV=development CLI_ONLY=true SB_TRAY=false nest start --watch",
@@ -36,11 +40,10 @@
     "start:app:prod": "cross-env SB_TRAY=false SB_USE_CWD=true node dist/main",
     "start:cli": "yarn start:app:cli",
     "start:cli:dev": "concurrently -k \"yarn start:app:cli:dev\"",
-    "start:dev": "concurrently -k \"yarn start:app:dev\" \"yarn start:tray\"",
-    "start:prod": "concurrently -k \"yarn start:app:prod\" \"yarn start:tray:prod\"",
+    "start:frontend": "cd frontend && yarn start",
+    "start:frontend:dev": "cd frontend && yarn start:dev",
     "start:tray": "cd tray && cross-env NODE_ENV=development yarn start",
     "start:tray:prod": "cd tray && cross-env SB_USE_CWD=true yarn start:prod",
-    "start": "concurrently -k \"yarn start:app\" \"yarn start:tray\"",
     "test": "jest",
     "test:cov": "jest --coverage",
     "test:debug": "node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand",
@@ -49,13 +52,6 @@
     "test:watch": "jest --watch"
   },
   "dependencies": {
-    "@fontsource/roboto": "4.5.0",
-    "@material-ui/core": "4.12.3",
-    "@material-ui/icons": "4.11.2",
-    "@material-ui/lab": "4.0.0-alpha.60",
-    "@material-ui/styles": "4.11.4",
-    "@mdi/js": "5.9.55",
-    "@mdi/react": "1.5.0",
     "@nestjs/common": "8.0.6",
     "@nestjs/core": "8.0.6",
     "@nestjs/platform-express": "8.0.6",
@@ -66,27 +62,15 @@
     "auto-launch": "5.0.5",
     "axios": "0.21.1",
     "better-sqlite3": "7.4.3",
-    "clsx": "1.1.1",
     "dotenv": "10.0.0",
     "execa": "5.1.1",
     "helmet": "4.6.0",
-    "lodash": "4.17.21",
     "loudness": "0.4.1",
-    "moment": "2.29.1",
     "mqtt": "4.2.8",
-    "next": "11.1.0",
     "node-notifier": "10.0.0",
     "open": "8.2.1",
     "peer": "0.6.1",
-    "peerjs": "1.3.2",
     "query-string": "7.0.1",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-json-view": "1.21.3",
-    "react-markdown": "7.0.1",
-    "react-moment": "1.1.1",
-    "react-syntax-highlighter": "15.4.4",
-    "rect-scaler": "1.0.0",
     "reflect-metadata": "0.1.13",
     "rimraf": "3.0.2",
     "rxjs": "7.3.0",
@@ -111,9 +95,6 @@
     "@types/mdns": "0.0.34",
     "@types/node": "16.7.10",
     "@types/node-notifier": "8.0.1",
-    "@types/react": "17.0.19",
-    "@types/react-dom": "17.0.9",
-    "@types/react-syntax-highlighter": "13.5.2",
     "@types/semver": "7.3.8",
     "@types/supertest": "2.0.11",
     "@types/uuid": "8.3.1",
@@ -165,7 +146,7 @@
   },
   "pkg": {
     "assets": [
-      ".next/**/*",
+      "frontend/.next/**/*",
       "node_modules/**/*",
       "public/**/*"
     ],
